requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.24.3
pytest==7.4.3
gtts==2.5.1
faster-whisper==1.0.3